fix(data-table): Fix UI freeze on big export (reproject geom/chunk with worker) Closes #2580
