+++ packages/geoview-core/src/core/components/data-table/json-export-button.tsx
@@ -1,15 +1,17 @@
-import { useCallback } from 'react';
+/* eslint-disable no-await-in-loop */
+import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import _ from 'lodash';
 
-import { Geometry, Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon } from 'ol/geom';
+// import { Geometry, Point, Polygon, LineString, MultiPoint } from 'ol/geom';
+// import { Geometry } from 'ol/geom';
 
 import { MenuItem } from '@/ui';
-import { logger } from '@/core/utils/logger';
-import { useMapStoreActions } from '@/core/stores/store-interface-and-intial-values/map-state';
-import { TypeJsonObject } from '@/core/types/global-types';
-import { useLayerStoreActions } from '@/core/stores/store-interface-and-intial-values/layer-state';
+// import { useMapStoreActions } from '@/core/stores/store-interface-and-intial-values/map-state';
 import { TypeFeatureInfoEntry } from '@/geo/map/map-schema-types';
+// import { TypeJsonObject } from '@/core/types/global-types';
+import { useLayerStoreActions } from '@/core/stores/store-interface-and-intial-values/layer-state';
+import { useAppStoreActions } from '@/core/stores/store-interface-and-intial-values/app-state';
+// import { logger } from '@/core/utils/logger';
 
 interface JSONExportButtonProps {
   rows: unknown[];
@@ -28,169 +30,37 @@ function JSONExportButton({ rows, features, layerPath }: JSONExportButtonProps):
   const { t } = useTranslation<string>();
 
   // get store value - projection config to transfer lat long and layer
-  const { transformPoints } = useMapStoreActions();
-  const { getLayer, queryLayerEsriDynamic } = useLayerStoreActions();
+  // const { transformPoints } = useMapStoreActions();
+  const { getLayer } = useLayerStoreActions();
 
   /**
    * Creates a geometry json
    * @param {Geometry} geometry - The geometry
    * @returns {TypeJsonObject} The geometry json
    */
-  const buildGeometry = useCallback(
-    (geometry: Geometry): TypeJsonObject => {
-      let builtGeometry = {};
+  // const buildGeometry = useCallback(
+  //   (geometry: Geometry): TypeJsonObject => {
+  //     let builtGeometry = {};
 
-      if (geometry instanceof Polygon) {
-        // coordinates are in the form of Coordinate[][]
-        builtGeometry = {
-          type: 'Polygon',
-          coordinates: geometry.getCoordinates().map((coords) => {
-            return coords.map((coord) => transformPoints([coord], 4326)[0]);
-          }),
-        };
-      } else if (geometry instanceof MultiPolygon) {
-        // coordinates are in the form of Coordinate[][][]
-        builtGeometry = {
-          type: 'MultiPolygon',
-          coordinates: geometry.getCoordinates().map((coords1) => {
-            return coords1.map((coords2) => {
-              return coords2.map((coord) => transformPoints([coord], 4326)[0]);
-            });
-          }),
-        };
-      } else if (geometry instanceof LineString) {
-        // coordinates are in the form of Coordinate[]
-        builtGeometry = { type: 'LineString', coordinates: geometry.getCoordinates().map((coord) => transformPoints([coord], 4326)[0]) };
-      } else if (geometry instanceof MultiLineString) {
-        // coordinates are in the form of Coordinate[][]
-        builtGeometry = {
-          type: 'MultiLineString',
-          coordinates: geometry.getCoordinates().map((coords) => {
-            return coords.map((coord) => transformPoints([coord], 4326)[0]);
-          }),
-        };
-      } else if (geometry instanceof Point) {
-        // coordinates are in the form of Coordinate
-        builtGeometry = { type: 'Point', coordinates: transformPoints([geometry.getCoordinates()], 4326)[0] };
-      } else if (geometry instanceof MultiPoint) {
-        // coordinates are in the form of Coordinate[]
-        builtGeometry = { type: 'MultiPoint', coordinates: geometry.getCoordinates().map((coord) => transformPoints([coord], 4326)[0]) };
-      }
+  //     if (geometry instanceof Polygon) {
+  //       builtGeometry = {
+  //         type: 'Polygon',
+  //         coordinates: geometry.getCoordinates().map((coords) => {
+  //           return coords.map((coord) => transformPoints([coord], 4326)[0]);
+  //         }),
+  //       };
+  //     } else if (geometry instanceof LineString) {
+  //       builtGeometry = { type: 'LineString', coordinates: geometry.getCoordinates().map((coord) => transformPoints([coord], 4326)[0]) };
+  //     } else if (geometry instanceof Point) {
+  //       builtGeometry = { type: 'Point', coordinates: transformPoints([geometry.getCoordinates()], 4326)[0] };
+  //     } else if (geometry instanceof MultiPoint) {
+  //       builtGeometry = { type: 'MultiPoint', coordinates: geometry.getCoordinates().map((coord) => transformPoints([coord], 4326)[0]) };
+  //     }
 
-      return builtGeometry;
-    },
-    [transformPoints]
-  );
-
-  /**
-   * Builds the JSON features section of the file
-   * @returns {string} Json file content as string
-   */
-  const getJsonFeatures = useCallback(
-    (theFeatures: TypeFeatureInfoEntry[]): TypeJsonObject[] => {
-      // Create GeoJSON feature
-      return theFeatures.map((feature) => {
-        const { geometry, fieldInfo } = feature;
-
-        // Format the feature info to extract only value and remove the geoviewID field
-        const formattedInfo: Record<string, unknown>[] = [];
-        Object.keys(fieldInfo).forEach((key) => {
-          if (key !== 'geoviewID') {
-            const tmpObj: Record<string, unknown> = {};
-            tmpObj[key] = fieldInfo[key]!.value;
-            formattedInfo.push(tmpObj);
-          }
-        });
-
-        return {
-          type: 'Feature',
-          geometry: buildGeometry(geometry?.getGeometry() as Geometry),
-          properties: formattedInfo,
-        } as unknown as TypeJsonObject;
-      });
-    },
-    [buildGeometry]
-  );
-
-  /**
-   * Builds the JSON file
-   * @returns {string} Json file content as string
-   */
-  const getJson = useCallback(
-    async (fetchGeometriesDuringProcess: boolean): Promise<string | undefined> => {
-      // Filter features from filtered rows
-      const rowsID = rows.map((row) => {
-        if (
-          typeof row === 'object' &&
-          row !== null &&
-          'geoviewID' in row &&
-          typeof row.geoviewID === 'object' &&
-          row.geoviewID !== null &&
-          'value' in row.geoviewID
-        ) {
-          return row.geoviewID.value;
-        }
-        return '';
-      });
-
-      const filteredFeatures = features.filter((feature) => rowsID.includes(feature.fieldInfo.geoviewID!.value));
-
-      // If must fetch the geometries during the process
-      if (fetchGeometriesDuringProcess) {
-        try {
-          // Split the array in arrays of 100 features maximum
-          const sublists = _.chunk(filteredFeatures, 100);
-
-          // For each sub list
-          const promises = sublists.map((sublist) => {
-            // Create a new promise that will resolved when features have been updated with their geometries
-            return new Promise<void>((resolve, reject) => {
-              // Get the ids
-              const objectids = sublist.map((record) => {
-                return record.geometry?.get('OBJECTID') as number;
-              });
-
-              // Query
-              queryLayerEsriDynamic(layerPath, objectids)
-                .then((results) => {
-                  // For each result
-                  results.forEach((result) => {
-                    // Filter
-                    const recFound = filteredFeatures.filter(
-                      (record) => record.geometry?.get('OBJECTID') === result.fieldInfo?.OBJECTID?.value
-                    );
-
-                    // If found it
-                    if (recFound && recFound.length === 1) {
-                      // Officially attribute the geometry to that particular record
-                      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                      (recFound[0].geometry as any).setGeometry(result.geometry);
-                    }
-                  });
-
-                  // Only now, resolve the promise
-                  resolve();
-                })
-                .catch(reject);
-            });
-          });
-
-          // Once all promises complete
-          await Promise.all(promises);
-        } catch (err) {
-          // Handle error
-          logger.logError('Failed to query the features to get their geometries. The output will not have the geometries.', err);
-        }
-      }
-
-      // Get the Json Features
-      const geoData = getJsonFeatures(filteredFeatures);
-
-      // Stringify with some indentation
-      return JSON.stringify({ type: 'FeatureCollection', features: geoData }, null, 2);
-    },
-    [layerPath, features, rows, getJsonFeatures, queryLayerEsriDynamic]
-  );
+  //     return builtGeometry;
+  //   },
+  //   [transformPoints]
+  // );
 
   /**
    * Exports the blob to a file
@@ -209,32 +79,147 @@ function JSONExportButton({ rows, features, layerPath }: JSONExportButtonProps):
     URL.revokeObjectURL(url);
   }, []);
 
-  /**
-   * Exports data table in csv format.
-   */
-  const handleExportData = useCallback((): void => {
-    const layer = getLayer(layerPath);
-    const layerIsEsriDynamic = layer?.type === 'esriDynamic';
+  // Helper function to serialize geometry
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const serializeGeometry = (geometry: any) => {
+    if (!geometry) return null;
+    return {
+      type: geometry.getType(),
+      coordinates: geometry.getCoordinates(),
+      // Add any other properties that might be needed
+    };
+  };
 
-    // Get the Json content for the layer
-    getJson(layerIsEsriDynamic)
-      .then((jsonString: string | undefined) => {
-        // If defined
-        if (jsonString) {
-          const blob = new Blob([jsonString], {
-            type: 'text/json',
-          });
+  const getJson = useCallback(
+    async function* (): AsyncGenerator<string> {
+      const worker = new Worker(new URL('./worker.ts', import.meta.url));
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const rowsIDSet = new Set(rows.map((row: any) => row?.geoviewID?.value).filter(Boolean));
+      const chunkSize = 1000; // Adjust based on performance testing
 
-          exportBlob(blob, `table-${layer?.layerName.replaceAll(' ', '-')}.json`);
+      // Get projection information
+      const projectionInfo = {
+        sourceCRS: 'EPSG:3978', // Projection.getSourceCRS(), // Replace with actual method to get source CRS
+        targetCRS: 'EPSG:4326', // Replace with actual method to get target CRS
+        // Add any other necessary projection parameters
+      };
+
+      // Initialize worker with projection info
+      worker.postMessage({ type: 'init', projectionInfo });
+
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const processChunk = async (chunk: any[], isFirst: boolean): Promise<string> => {
+        return new Promise((resolve) => {
+          worker.onmessage = (event) => resolve(event.data);
+          worker.postMessage({ type: 'process', chunk, isFirst });
+        });
+      };
+
+      try {
+        for (let i = 0; i < features.length; i += chunkSize) {
+          const chunk = features.slice(i, i + chunkSize);
+          const serializedChunk = chunk
+            .filter((feature) => rowsIDSet.has(feature.fieldInfo.geoviewID?.value))
+            .map((feature) => ({
+              geometry: serializeGeometry(feature.geometry?.getGeometry()?.clone()),
+              properties: Object.fromEntries(
+                Object.entries(feature.fieldInfo)
+                  .filter(([key]) => key !== 'geoviewID')
+                  .map(([key, value]) => [key, value?.value])
+              ),
+            }));
+
+          if (serializedChunk.length > 0) {
+            const result = await processChunk(serializedChunk, i === 0);
+            yield result;
+          }
+
+          // Allow UI to update
+          // eslint-disable-next-line no-promise-executor-return
+          await new Promise((resolve) => setTimeout(resolve, 0));
         }
-      })
-      .catch((err) => {
-        // Log
-        logger.logPromiseFailed('Not able to export', err);
-      });
-  }, [exportBlob, getJson, getLayer, layerPath]);
 
-  return <MenuItem onClick={handleExportData}>{t('dataTable.jsonExportBtn')}</MenuItem>;
+        yield ']}';
+      } finally {
+        worker.terminate();
+      }
+    },
+    [features, rows]
+  );
+
+  const [isExporting, setIsExporting] = useState(false);
+  const { addMessage } = useAppStoreActions();
+
+  const handleExportData = useCallback(async () => {
+    setIsExporting(true);
+    try {
+      const jsonGenerator = getJson();
+      const chunks = [];
+      let i = 0;
+
+      for await (const chunk of jsonGenerator) {
+        chunks.push(chunk);
+        i++;
+
+        // Optionally update progress here
+        addMessage('info', `Processing ${i * 1000} ... of ${rows.length}`);
+      }
+
+      const fullJson = chunks.join('');
+      const blob = new Blob([fullJson], { type: 'application/json' });
+      exportBlob(blob, `table-${getLayer(layerPath)?.layerName.replaceAll(' ', '-')}.json`);
+      // Do something with the full JSON, e.g., save to file
+    } catch (error) {
+      console.error('Export failed:', error);
+    } finally {
+      setIsExporting(false);
+    }
+  }, [getJson]);
+
+  // const handleExportData = useCallback(async (): Promise<void> => {
+  //   try {
+  //     const jsonGenerator = getJson();
+
+  //     const stream = new ReadableStream({
+  //       async start(controller) {
+  //         try {
+  //           for await (const chunk of jsonGenerator) {
+  //             controller.enqueue(new TextEncoder().encode(chunk));
+  //           }
+  //         } catch (error) {
+  //           console.error('Error in stream processing:', error);
+  //           controller.error(error);
+  //         } finally {
+  //           controller.close();
+  //         }
+  //       },
+  //     });
+
+  //     const response = new Response(stream);
+  //     const blob = await response.blob();
+
+  //     exportBlob(blob, `table-${getLayer(layerPath)?.layerName.replaceAll(' ', '-')}.json`);
+  //   } catch (error) {
+  //     console.error('Export failed:', error);
+  //     // Optionally, display an error message to the user
+  //   }
+  // }, [exportBlob, getJson, getLayer, layerPath]);
+
+  // const handleExportData = useCallback((): void => {
+  //   try {
+  //     const jsonString = getJson();
+  //     const blob = new Blob([jsonString], { type: 'application/json' });
+  //     exportBlob(blob, `table-${getLayer(layerPath)?.layerName.replaceAll(' ', '-')}.json`);
+  //   } catch (error) {
+  //     logger.logError('Table download as GeoJSON failed:', error);
+  //   }
+  // }, [exportBlob, getJson, getLayer, layerPath]);
+
+  return (
+    <MenuItem onClick={handleExportData} disabled={isExporting}>
+      {t('dataTable.jsonExportBtn')}
+    </MenuItem>
+  );
 }
 
 export default JSONExportButton;
