+++ packages/geoview-core/src/geo/utils/renderer/geoview-renderer.ts
@@ -1220,6 +1220,7 @@ export async function getLegendStyles(styleConfig: TypeStyleConfig | undefined):
       legendStyles.Polygon = layerStyles.Polygon;
     }
     return legendStyles;
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (error) {
     return {};
   }
